# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "img2ec",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/projects/{project_id}/skus/{sku_id}/download-bundle": {
      "post": {
        "tags": [
          "outputs"
        ],
        "summary": "Download Bundle",
        "description": "打包用户选定的资料：主图/ + SKU图/ + 详情图/ + 文案.txt",
        "operationId": "download_bundle_api_projects__project_id__skus__sku_id__download_bundle_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "sku_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sku Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BundleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/skus/{sku_id}/download-bundle-all": {
      "post": {
        "tags": [
          "outputs"
        ],
        "summary": "Download Bundle All",
        "description": "打包用户当前变体的全平台资料。结构：\n主图/        ← 跨平台共用（cur.main）\nSKU图/       ← 跨平台共用（variant.sku_thumb_paths）\n<平台>/详情图/detail-template.jpg\n<平台>/文案.txt\nREADME.txt",
        "operationId": "download_bundle_all_api_projects__project_id__skus__sku_id__download_bundle_all_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "sku_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sku Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BundleAllRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/skus/{sku_id}/download": {
      "get": {
        "tags": [
          "outputs"
        ],
        "summary": "Download Sku Zip",
        "operationId": "download_sku_zip_api_skus__sku_id__download_get",
        "parameters": [
          {
            "name": "sku_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sku Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/download-all": {
      "get": {
        "tags": [
          "outputs"
        ],
        "summary": "Download Project Zip",
        "description": "打包所有 done 状态 SKU 的 outputs/，每个 SKU 一个目录。",
        "operationId": "download_project_zip_api_projects__project_id__download_all_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BundleAllRequest": {
        "properties": {
          "variant_id": {
            "type": "string",
            "title": "Variant Id"
          },
          "main_keys": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Main Keys",
            "default": []
          },
          "detail_keys": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Detail Keys",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "variant_id"
        ],
        "title": "BundleAllRequest"
      },
      "BundleRequest": {
        "properties": {
          "platform": {
            "type": "string",
            "title": "Platform"
          },
          "variant_id": {
            "type": "string",
            "title": "Variant Id"
          },
          "main_keys": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Main Keys",
            "default": []
          },
          "detail_keys": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Detail Keys",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "platform",
          "variant_id"
        ],
        "title": "BundleRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}